+++ functions.php
@@ -15,6 +15,7 @@ require_once get_template_directory() . '/inc/authentication.php';
 require_once get_template_directory() . '/inc/search.php';
 require_once get_template_directory() . '/inc/posts.php';
 require_once get_template_directory() . '/inc/volumes.php';
+require_once get_template_directory() . '/inc/query-sort.php';
 require_once get_template_directory() . '/inc/tags.php';
 require_once get_template_directory() . '/inc/more-link.php';
 require_once get_template_directory() . '/inc/cz-continue-reading.php';