+++ pages/page-volumi.php
@@ -1,15 +1,43 @@
 <section class="volumes-list">
-	<h1>Tutti i Volumi</h1>
 	<?php
-	$volumes_query = new WP_Query(
-		array(
-			'post_type'      => 'volume',
-			'post_status'    => 'publish',
-			'posts_per_page' => -1,
-			'orderby'        => 'title',
-			'order'          => 'ASC',
-		)
-	);
+	$sort_fields = [
+		'date'     => __( 'Data di pubblicazione', 'textdomain' ),
+		'modified' => __( 'Data di modifica', 'textdomain' ),
+		'title'    => __( 'Titolo', 'textdomain' ),
+		'author'   => __( 'Autore', 'textdomain' ),
+	];
+	$sort_default_order = [
+		'date'     => 'DESC',
+		'modified' => 'DESC',
+		'title'    => 'ASC',
+		'author'   => 'ASC',
+	];
+	[ $orderby_param, $order_param ] = cz_resolve_sort_params( $sort_fields, $sort_default_order );
+
+	$volumes_args = [
+		'post_type'      => 'volume',
+		'post_status'    => 'publish',
+		'posts_per_page' => -1,
+	];
+
+	if ( 'author' === $orderby_param ) {
+		$volumes_query = cz_query_sorted_by_author_name( $volumes_args, $order_param );
+	} else {
+		$orderby_map = [
+			'date'     => 'date',
+			'modified' => 'modified',
+			'title'    => 'title',
+		];
+		$volumes_args['orderby'] = [ $orderby_map[ $orderby_param ] => $order_param, 'ID' => $order_param ];
+		$volumes_query = new WP_Query( $volumes_args );
+	}
+
+	?>
+	<header class="archive-header">
+	  <h1 class="archive-title">Tutti i Volumi</h1>
+	  <?php cz_render_sort_control( $sort_fields, $orderby_param, $order_param ); ?>
+	</header>
+	<?php
 
 	if ( $volumes_query->have_posts() ) :
 		echo '<ul class="volumes-items">';